refactor: Player 인터페이스와 구현체 User 추가
Player를 인터페이스로 만들고 새로 구현체 User 클래스를 만듦. Computer 클래스 역시 Player의 구현체가 되도록 수정
